+++ docs/changelog.md
@@ -2,8 +2,8 @@
 
 ## "version": "0.0.0"
 
-- srcSet
-- tags for individual picture
+- fixed the common images on-change loading message
+- srcset and tags for individual picture
 - make title/link input fields on the create post section fill their containers
 - auth, layout, material and product libs based on the Duncan Hunter course
 - pictures, trends and trendy-layout created for the trend topic post creation process